+++ src/wtp_game.cpp
@@ -1929,6 +1929,7 @@ void computeBase(int baseId, bool resetWorkedTiles)
 	{
 		Bases[baseId].worked_tiles = 0;
 		Bases[baseId].specialist_total = 0;
+		Bases[baseId].specialist_adjust = 0;
 	}
 	
 	set_base(baseId);